+++ NEWS.md
@@ -1,5 +1,10 @@
 # News
 
+## 0.1.1
+
+- Added `get_goi_counts()` to `R/qc_helpers.R` for counting expression of
+  genes of interest (e.g. viral genes) in a selected cell population.
+
 ## 0.1.0
 
 - Initial extraction of reusable helpers into `baia.seurat.helpers`.